+++ glotaran/project/project_data_registry.py
@@ -32,7 +32,7 @@ class ProjectDataRegistry(ProjectRegistry):
     def import_data(
         self,
         dataset: str | Path | xr.Dataset | xr.DataArray,
-        name: str | None = None,
+        dataset_name: str | None = None,
         allow_overwrite: bool = False,
         ignore_existing: bool = False,
     ):
@@ -42,7 +42,7 @@ class ProjectDataRegistry(ProjectRegistry):
         ----------
         dataset : str | Path | xr.Dataset | xr.DataArray
             Dataset instance or path to a dataset.
-        name : str | None
+        dataset_name : str | None
             The name of the dataset (needs to be provided when dataset is an xarray instance).
             Defaults to None.
         allow_overwrite: bool
@@ -55,10 +55,10 @@ class ProjectDataRegistry(ProjectRegistry):
         ValueError
             When importing from xarray object and not providing a name.
         """
-        if isinstance(dataset, (xr.DataArray, xr.Dataset)) and name is None:
+        if isinstance(dataset, (xr.DataArray, xr.Dataset)) and dataset_name is None:
             raise ValueError(
                 "When importing data from a 'xarray.Dataset' or 'xarray.DataArray' "
-                "it is required to also pass a name."
+                "it is required to also pass a ``dataset_name``."
             )
         if isinstance(dataset, xr.DataArray):
             dataset = dataset.to_dataset(name="data")
@@ -69,10 +69,10 @@ class ProjectDataRegistry(ProjectRegistry):
             if dataset.is_absolute() is False:
                 dataset = (self.directory.parent / dataset).resolve()
 
-            name = name or dataset.stem
+            dataset_name = dataset_name or dataset.stem
             dataset = load_dataset(dataset)
 
-        data_path = self.directory / f"{name}.nc"
+        data_path = self.directory / f"{dataset_name}.nc"
         if data_path.exists() and ignore_existing:
             return
         save_dataset(dataset, data_path, allow_overwrite=allow_overwrite)